requirements.txt

streamlit
openai
serpapi
pytesseract
Pillow
